+++ CHANGELOG.md
@@ -1,1 +1,10 @@
+## [1.1.0]
+- Added customization for animation on text.
+- Updated example with animations.
+
+## [1.0.1]
+- Added License.
+- Updated example.
+
 ## [1.0.0]
+Initial Release.